+++ src/algorithms/rem_tree.tpp
@@ -20,7 +20,7 @@ have not been implemented as expected.
 template <typename T, typename U>
 vector<T> remainder_tree_basic(const vector<T>& A, const vector<U>& m, const T& V) {
 	assert (A.size() == m.size());
-	long long N = A.size();
+	long N = A.size();
 
 	vector<U> mTree = subproduct_tree<U> (m, 1); //Default value of P corresponds to root node, but passing in just to be safe
 	vector<T> ATree = subproduct_tree_askew<T,U> (A, mTree, 1);
@@ -28,14 +28,14 @@ vector<T> remainder_tree_basic(const vector<T>& A, const vector<U>& m, const T& 
 	vector<T> CTree(2*N);
 
 	CTree[1] = V*A[0]; //recall the tree is 1-indexed
-	for (long long i = 1; i < N; ++i) { //only go up to second-lowest layer
+	for (long i = 1; i < N; ++i) { //only go up to second-lowest layer
 		CTree[2*i] = CTree[i]%mTree[2*i];
 		//CTree[2*i + 1] = (CTree[i]*ATree[2*i + 1])%mTree[2*i + 1];
 		
 		CTree[2*i + 1] = CTree[i];
 		CTree[2*i + 1].mulmod(ATree[2*i + 1], mTree[2*i + 1]);
 
-		//Always kill the parents when no long longer needed
+		//Always kill the parents when no longer needed
 		CTree[i] = T();
 		ATree[i] = T();
 		mTree[i] = U();
@@ -56,11 +56,11 @@ vector<T> remainder_tree_basic(const vector<T>& A, const vector<U>& m, const T& 
 //Because of this, try to use powers of 2 for best performance.
 template <typename T, typename U>
 vector<T> remainder_tree_padded(vector<T> A, vector<U> m, const T& V) {
-	long long N = A.size();
+	long N = A.size();
 	assert (A.size() == m.size());
 
-	long long height = ceil(log2(N-1));
-	long long N_pad = (1<<height);
+	long height = ceil(log2(N-1));
+	long N_pad = (1<<height);
 
 	A.resize(N_pad, 0);
 	m.resize(N_pad, 1);
